Refactored IndicatorLike to IndicatorBase.
