+++ ts/Stock/Indicators/AD/ADIndicator.ts
@@ -18,7 +18,7 @@ import type {
     ADParamsOptions
 } from './ADOptions';
 import type ADPoint from './ADPoint';
-import type { IndicatorLinkedSeriesBase } from '../IndicatorLike';
+import type { IndicatorLinkedSeriesBase } from '../IndicatorBase';
 import type IndicatorValuesObject from '../IndicatorValuesObject';
 import type LineSeries from '../../../Series/Line/LineSeries';
