+++ ts/Stock/Indicators/ATR/ATRIndicator.ts
@@ -19,7 +19,7 @@ import type {
     ATRParamsOptions
 } from './ATROptions';
 import type ATRPoint from './ATRPoint';
-import type { IndicatorLinkedSeriesBase } from '../IndicatorLike';
+import type { IndicatorLinkedSeriesBase } from '../IndicatorBase';
 import type IndicatorValuesObject from '../IndicatorValuesObject';
 import type LineSeries from '../../../Series/Line/LineSeries';
