+++ ts/Stock/Indicators/DPO/DPOIndicator.ts
@@ -19,7 +19,7 @@ import type {
     DPOParamsOptions
 } from './DPOOptions';
 import type DPOPoint from './DPOPoint';
-import type { IndicatorLinkedSeriesBase } from '../IndicatorLike';
+import type { IndicatorLinkedSeriesBase } from '../IndicatorBase';
 import type IndicatorValuesObject from '../IndicatorValuesObject';
 import type LineSeries from '../../../Series/Line/LineSeries';
