+++ ts/Stock/Indicators/MACD/MACDIndicator.ts
@@ -16,7 +16,7 @@
 
 import type ColorString from '../../../Core/Color/ColorString';
 import type ColumnSeries from '../../../Series/Column/ColumnSeries';
-import type { IndicatorLinkedSeriesBase } from '../IndicatorLike';
+import type { IndicatorLinkedSeriesBase } from '../IndicatorBase';
 import type IndicatorValuesObject from '../IndicatorValuesObject';
 import type LineSeries from '../../../Series/Line/LineSeries';
 import type {
