+++ ts/Stock/Indicators/MFI/MFIIndicator.ts
@@ -18,7 +18,7 @@
  *
  * */
 
-import type { IndicatorLinkedSeriesBase } from '../IndicatorLike';
+import type { IndicatorLinkedSeriesBase } from '../IndicatorBase';
 import type IndicatorValuesObject from '../IndicatorValuesObject';
 import type LineSeries from '../../../Series/Line/LineSeries';
 import type {
